Skip building anything besides "Docs" on "docs" branches.
This should avoid needing to include `[skip ci]` on every commit on branches
which are committed to docs, though will still provide us one "master"
branch run of tests when the PR lands.

We may want to remove this (maybe? longer-term), but this seems like a
short-term "win" on this branch for the time being.

cc: @jbaxleyiii @evans