Introduce a `netlify.toml` configuration to fix docs generation on Netlify.
Without this file, Netlify was choosing to build the documentation using the
npm scripts present in the root `package.json`, rather than the `package.json`
in the `docs/` directory, as it should be.

While it's strange that this behavior just cropped up around a recent swap to
using `@apollographql/apollo-upload-server`, I suspect that the problem is
more deeply rooted than that and we were just accidentally getting away with
the past behavior - which was defined less explicitly in the Netlify UI.

Now, with this `netlify.toml` file, we can run the exact build step we need
for documentation and also gain the ability to have per-branch control over
that configuration, which has been necessary in the past.