Add a Codecov configuration file to adjust the defaults.
In particular, we're not 100% sold on the Codecov charts - which add some
noise, but seem like they could be helpful at higher coverage percentages.

Also, for the time-being, we'll lower the target for `patch`s (i.e. PRs) to
zero, just to see how the values get reported in the CI checks.

Ref: https://docs.codecov.io/docs/commit-status