Don't enable uploads when using an unsupported Node.js in tests.
Now that there are specific versions of Node.js which don't support file
uploads (namely, <= 8.5.0) we need to explicitly disable uploads on those
versions, similar to how those users must opt-in to that behavior by setting
`uploads: false` in their Apollo Server constructor options.

This effectively accomplishes exactly that, but only when necessary.