(op-reg) Forbid operations when `forbidUnregisteredOperations` is enabled.
This re-lands 956c8f02bc913afe355f40e9179041167a9b8267 which was a
work-in-progress commit which added logging at the correct places, but
stopped short of actual enforcement.

Prior to this new functionality, unregistered operations were forbidden by
default, even if the `forbidUnregisteredOperations` variable which this
commit (now) implements was `true` or not.

The documentation for the operation registry plugin has always stated that
this parameter must be set in order to enable this behavior, so this shouldn't
be interpreted as a breaking change - particularly since the previous
behavior was an `alpha` release on a pre-1.x version.

This commit makes the implementation more flexible by allowing incremental
adoption of operation safe-listing since it allows the plugin to be
installed and the enforcement operation to be observed via logging
(immediately) and hopefully soon via Apollo Engine.

Finally, for servers that need to selectively enable enforcement, the
`forbidUnregisteredOperations` parameter accepts a predicate function which
receives the request's `GraphQLRequestContext` as the first parameter.  This
will allow implementers to selectively disable the enforcement when, for
example, a header is present that identifies the requester as a "super-user"
(or other authorized user who is not subject to operation safe-listing).

For example:

```
    require("apollo-server-plugin-operation-registry")({
      // De-structure the request context to get the `headers` map.
      forbidUnregisteredOperations({ request: { http: { headers } } }) {
        // If the magic header is in place, allow any unregistered operation.
        // Of course, additional validation is possible!
        if (headers.get('x-let-me-in')) {
          return false;
        }
        // Otherwise, enforce it.
        return true;
      }
    })
```