Compare Engine reporting's `privateHeaders` case-insensitively, as documented.
The documentation for `privateHeaders`[[0]] suggests that it is
case-insensitive.  While that statement is true, and the incoming header is
lower-cased before checking it against the `privateHeaders` configuration,
it assumed that the headers in the `privateHeaders` object were specified in
lower-case.

This changes the comparison to lower-case both sides prior to determining
equality.

[0]: https://github.com/apollographql/apollo-server/blob/27323f4d/packages/apollo-engine-reporting/src/agent.ts#L67-L70

Fixes: https://github.com/apollographql/apollo-server/issues/2273