TEMP: Revert "feat(Fastify) Apollo server Fastify integration #626 (#1971)"
This TEMPORARILY reverts commit 75ce372dffc17256cc31e3a847b6c9a2f020f59a,
which was the result of the work done in #1971 by @rkorrelboom.

Unfortunately, we need to put this on ice while we wait for movement on a
package naming conflict.  The dialog surrounding this is under way, as
I've explained in the PR:

https://github.com/apollographql/apollo-server/pull/1971#issuecomment-456817749

I'm excited to re-land this in an upcoming version, but there's no reason to
block the 2.4.0 release for it right now.

I will open a new PR with the work from #1971 in due time.