Switch `apollo-graphql` path-based import to use `lib` rather than `dist`. (#23)
Since we've moved the `apollo-graphql` package from `apollo-server` to
`apollo-tooling`, it's now subjected to the new TypeScript compilation rules
for that repository.

(See: https://github.com/apollographql/apollo-tooling/pull/1018)

Those compilation rules differ slightly and put their output in the `lib`
directory rather than the `dist` directory.  In an ideal world, we'd fix
that at publish time so the actual `outDir` didn't matter (i.e. we could
actually publish from the `dist/` directory, something which `lerna`
actually supports!), but for now, changing this import should be good.

I'd intentionally not done this before (and instead just pinned the version
exactly to a version which I knew was still utilizing the previous `dist/`
compilation `outDir`) as we'd been waiting for the first publish to occur
from that repository, but that publish has now happened and Renovate has
come through and updated it so it's time to make the change more
permanently!

Follows-up: https://github.com/apollographql/apollo-platform-commercial/pull/10