Pass `'queryPlan.hideQueryPlanResponse'` to conform with new `ISettings`.
It appears that the `hideQueryPlanResponse` is now a required member of the
`ISettings` interface within `graphql-playground-html`, thereby mandating
that any existing implementer of `ISettings` (via, for example,
`MiddlewareOptions` or `RenderPageOptions` respective `settings` properties
woudl necessitate that they have that property added.

This might be possible if `'queryPlan.hideQueryPlanResponse'` was not a
required property, but merely optional?

cc @trevor-scheer