Revert "chore(deps): update dependency lerna to v3.15.0 (#2810)" and pin.
In the same spirit as 6436103165ac21592cd6559050cdf0368aae61d3, this reverts
commit 6b311dac2d1e2f6e36815fff70f66603c56f7c9f which Renovate re-upgraded
in #2810.

Therefore, this commit also puts a restriction within Renovate's
configuration to prevent further upgrades.  Unfortunately, this means we're
stuck at this version of Lerna until we get time to make a reproduction for
the issue I logged when this first cropped up in
https://github.com/lerna/lerna/issues/2100, but I haven't yet found the
cycles to make that happen.

I guess our repository is a reliable reproduction, right now!