Rename confusing variable
This variable was named `forbidUnregisteredOperation`, but it could
actually be the result of a custom predicate function that could make up
its own mind about whether an operation should be forbidden. Moreover,
the behavior of the plugin depends not only on this rather prescriptive
sounding variable, but also on whether we are running in dryRun mode or
not.