Start running tests on Node.js v12. (#2983)
* Start running tests on Node.js v12.

While not currently a so-called "Long-Term-Support" (LTS) release, Node.js
v12 is on target to become LTS in October 2019 and Node.js 8 will end LTS
earlier than previous even-numbered Node.js releases.

See the Node.js schedule: https://github.com/nodejs/Release#release-schedule

With that only a few months away, it's time to start getting a read on
whether we're going to pass tests in that new engine.

* Remove `meteor-promise` and `fibers`, which appear to be unused entirely.

While originally introduced in
https://github.com/apollographql/apollo-server/pull/92 to accommodate
Meteor, it seems that we're no longer making `Promise`s Fiber-aware using
the `meteor-promise` library which leverages `fibers` under-the-hood.

Since this package seems to be the sole reason that bootstrapping is failing
on Node.js 12.  To support Node.js 12, we could update to `fibers@4` which
accommodates changes to the V8 API which have been made in recent Node.js
versions, the `fibers@4` implementation is only compatible with Node.js 10
and 12 and is incompatible with Node.js 8, which we still actively support
(though not for much longer!).