[Docs] [apollo-server-lambda] document usage of playground and defaultPlaygroundOptions (#2103)
* document usage of playground and defaultPlaygroundOptions

i was trying out @stubailo's tutorial here https://blog.apollographql.com/deploy-a-fullstack-apollo-app-with-netlify-45a7dfd51b0b and noticed that the playground no longer shows up when hitting the raw endpoint. this must have been a breaking change made somewhere, I can't track it down. it took a while to figure out what the right config was, which seems to be passing `defaultPlaygroundOptions`. I'm hoping to update the docs so future me's wont wonder about this.

* Update README.md

* Remove import of `defaultPlaygroundOptions`.