feat: Add support for value types in federation (#3063)
Value types are type definitions that live on multiple services (more specifically - types, inputs, interfaces, and unions). We've seen the desire internally and within the community to support the idea of a common type that multiple services can reference. A value type is identical by name, kind, and fields across all services.

In order to accomplish this, we've extended two of the existing validators: one for type names, and the other for field names. The changes to the validators permit duplicate type names and fields in the case that we've encountered a value type. The validators also provide some actionable, user-friendly errors in the case that we encounter a kind or field type mismatch.