Revert "Update dependencies accept & boom to new namespace (#30… (#3111)
This reverts commits 2810e51b, ed2aff5c and 3847ec41.

While the change in #3089 change was seemingly justified and straightforward
(and as #3106 points out, perhaps _too_ straightforward!  :smile:), it seems
that there was more to the change from `accept` to `@hapi/accept` and `boom`
to `@hapi/boom` than just a name change.  It seems that the packages may
have dropped support for Node.js 6.

While Node.js 6 is no longer supported by the Node Foundation, for
semantic versioning reasons we weren't intending to completely drop support
for it until Apollo Server 3.x.

Maybe there's a more granular update here that maintains this support, but
this update wasn't as simple as it would have seemed to be at face value
(i.e. not a straight package rename of an otherwise internal dependency).

Happy to consider re-landing a similar thing, but we need to make it more
clear in the `CHANGELOG.md` and possibly reserve it for Apollo Server 3.x,
but for now, should revert this to make sure we're not shipping a broken
integration for Hapi users, as reported in #3106.