Remove destructive code formatting tooling.
This change should not stop any developers from using auto code formatting
tools (e.g. Prettier, which was in place prior to this commit).

Developers who would like to keep using such tools merely need to highlight
their changes and choose the "Format Selection" (or similar) option in their
editor.  If they choose to use Prettier, no configuration file is necessary;
this project is accepting of the default Prettier settings, so long as they
aren't used to reformat entire files/code-bases.  For more details, please
read this incredibly long commit message.  I hope it'll help highlight
another side to automated, enforced, opinionated code-formatting tools. <3

Code formatting is, as I think this long commit message will continue to
point out, a subjective preference.  Code formatting tools, like Prettier,
have tried to end formatting disagreements by picking a single set of
"unwavering" rules.  It's hard to win at this, but as long as I'm involved
with this project, I will not be nit-picking anyone's choice of code-style
so long as it is not intentionally trying to be unnecessarily unorthodox.
Also, even if it is, in my opinion, appearing to be unnecessarily unorthodox
I pledge to try and understand and be empathetic to the reasoning for it.  I
would encourage others to evaluate, discuss, and react similarly!

At a high-level, my preference is: Try to match surrounding style, leave
comments when necessary, try to think about how to best represent your
change itself and consider the future of the code — at least briefly.

That said, the formatting of code isn't always always best determined by an
opinionated tool.  On more than one occasion, Prettier has made decisions
for this project which have resulted in decreased code clarity in our
changesets (i.e. pull-requests; PRs).

For example, even removing a single character from a variable can result in
a changeset with anywhere from ~2 to several-hundred of lines of unnecessary
changes.  As a very small example of this, consider [4bd1f901db][1], a
change to a misspelled variable, which resulted in the Prettification seen
in [d6b78d0cb93940][2].

Now amplify that change to a large function where the function parameters
changed by plus or minute two characters and the hanging indent
unnecessarily changes so that it can fit more things on the previous line.

When I first ran into the spelling example above, I was bothered and again
tempted to remove Prettier.  I knew that I had previous examples of this
amplified change being problematic, but couldn't quickly dig up an example
to make a case.  Low and behold, it only took five days before another
example came up!:

See the relatively simple change in [f343e323386c][3], which stats out to:

    packages/apollo-server-koa/src/__tests__/ApolloServer.test.ts | 22 ++++++++++++++++++++--
     1 file changed, 20 insertions(+), 2 deletions(-)

Now see what Prettier does to this brief change in [da3f5c3210a99][4]:

    packages/apollo-server-koa/src/__tests__/ApolloServer.test.ts | 1124 +++++++++++++++++++-------------------
     1 file changed, 562 insertions(+), 562 deletions(-)

With Prettier (or most opinionated automatic code formatting tools), the
entire function must now be reformatted when the clarity of the function was
not at all in jeopardy!

Not only is this unnecessarily changing code, but it also buries "blame" for
these changes across an extra ~500 lines of code through repeated rounds of
similar re-linting.  In my opinion, those additional lines being changed
aren't at all useful for the digestibility of a project.  When I open a PR,
for the success of the code, I prefer that the changes not be clouded with
artifacts that make it more difficult for the reviewer to understand.
Conversely, when I'm reviewing a PR, I prefer to avoid the same.  Unrelated
formatting changes — automatic or not — are the most frequent contributors
to this pain, in my experience.

As lesser, but still related points: It's not at all uncommon for
"opinionated" tools to change their opinions.  On more than one occasion,
I've updated a dependency in this repository only to have to re-format
swaths of code:

- https://github.com/apollographql/apollo-server/pull/846
- https://github.com/apollographql/apollo-server/pull/2203
- https://github.com/apollographql/apollo-server/pull/2572

Plus, on other occasions, I've had to wade through over-complicated merge
conflicts between branches when otherwise capable (Git) merge strategies
were sandbagged by formatting differences.

All this said, I want to emphasize that I really encourage anyone to use
tools that facilitate the automatic formatting of their code.  I just ask
that they can do it with a limited scope that merely touches the code
they're changing (e.g. "Format Selection").  If there is no way to avoid it
and it's absolutely necessary to re-format hundreds of lines of code, that's
okay!  There are certainly cases where very large reformatting is warranted,
but it seems better if it's the exception rather than the norm.

I hope that there's an understanding that there are times that tools that
work for one developer/project/concept, it won't work for another.  As a
regular contributor to this project, I personally find that Prettier
frustrates me more often than it helps me (long-term and short-term
considerations both being had).

Conversely, I know that Prettier greatly simplifies the workflow for many!
I love that!  But while someone might really enjoy `editor.formatOnSave`
re-formatting their code, imagine my own surprise when it does exactly that.

Dialogue and conversation is very welcome here.  I certainly don't want to
push anyone away from contributing, but I need to also be realistic about my
own frustrations.  I'm excited for tooling (which apparently doesn't exist
yet!) that can help improve this for those on either side of the Prettier
picket fence, but until that time comes, I think Prettier will better left
to the concern of the developers who wish to employ it.

[1]: https://github.com/apollographql/apollo-server/commit/4bd1f901dbeceb69638703cfa724576666fc2e31
[2]: https://github.com/apollographql/apollo-server/commit/d6b78d0cb93940675b193627695b04d2ceb59ae7
[3]: https://github.com/apollographql/apollo-server/commit/f343e323386c3f246995b540f17a8ac6fa2aa99c
[4]: https://github.com/apollographql/apollo-server/commit/da3f5c3210a99214464476012c2b1d276416a985