Remove memory leak from guaranteed unresolved/unrejected Promise.
I question whether or not we want to actually guarantee async execution of
this method and not give the opportunity to the user to actually make it
block if they'd like to, but we certainly should ensure that Promises should
not be created on every request and never resolve or reject, which seems
like a prime avenue for a DoS via memory leak.

I'm committing this because I stumbled across it and I feared it getting out,
but would appreciate eyes on it from @evans, if he has the time!