Gateway cached DataSource bug (#3412)
This fixes a bug in the gateway that prevents it from constructing
new `DataSource`s when composition updates and the url of an
underlying service changes. Currently the creation of a `DataSource`
is cached based on its `serviceName`, but really the caching should
depend on both the `serviceName` and `url` of the `ServiceDefinition`.