Reduce height of tall conditionals by returning early.
At this stage, we're in a mode of enforcement only.  Dry-run is ruled out and
there is no chance of a predicate decision changing.

Therefore, some slight tweaks in the ordering of operations here should make
the code easier to reason about since it avoids the need to remember how many
forks of logic we're inside.

To summarize this change which Git will make difficult to read:

This change moves the `logger.debug` for `dryRun` mode above the actual
enforcement (i.e. non-`dryRun`!) section and have it `return`, rather than
having a tall `if (!options.dryRun)` with many lines of logic when there are
no remaining forks of logic in the function after it that need to be
considered.