Use the typings from `Options` to resolve `any`-typed callbacks in tests.
The `onUnregisteredOptions` and `onForbiddenOperation` hooks weren't
properly typed in the mocked tests, which was allowing them to have the
incorrect type arguments.  This wouldn't be a problem with in-lined
definitions of these options - since that would be inferred from the parent
object literal's definition, but for tests where we assign the function to a
variable first, it is problematic.