To allow tests to shutdown quickly, use `sendReportsImmediately`.
Since the operation registry is now using the `engine` options from the
`ApolloServer` constructor in order to get its API key, we are effectively
instantiating Engine Reporting and its agent.  Since we're running in a test
environment, those transmissions are fake, but without setting this option,
they are still trying to be batch transmitted at a preset interval.  In
the Jest test environment, this just causes the tests to not finish as
quickly as Jest expects them to, and it spews warnings about open file
handles.

This should make that slightly better, though it would be ideal if we never
actually instantiated engine agent for reporting purposes (though we want it
for `apiKeyHash` receiving purposes.